bike-store/backend/requirements.txt
flask==3.0.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
python-dotenv==1.0.0
gunicorn==21.2.0
